feat: update server port to 55001 and refactor environment imports
