--- src/config/env.ts
@@ -1,2 +1,0 @@
-export { serverEnv } from "./env.server";
-export { clientEnv } from "./env.client";
