+++ src/features/tasks/server.ts
@@ -1,6 +1,6 @@
 import { createServerFn } from "@tanstack/react-start";
 import axios from "redaxios";
-import { serverEnv } from "~/config/env";
+import { serverEnv } from "~/config/env.server";
 import { Task } from "./types";
 import z, { date } from "zod";
