+++ vite.config.ts
@@ -8,7 +8,8 @@ import tailwindcss from "@tailwindcss/vite";
 
 export default defineConfig({
   server: {
-    port: 3000,
+    port: 55001,
+    host: true,
   },
   plugins: [
     tsConfigPaths({
